+++ modules/jaxb-xml-binding/geotk-xml-wms/src/main/java/org/geotoolkit/wms/xml/v130/LogoURL.java
@@ -16,7 +16,6 @@
  */
 package org.geotoolkit.wms.xml.v130;
 
-import java.math.BigInteger;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlAttribute;
@@ -63,10 +62,10 @@ public class LogoURL {
     private OnlineResource onlineResource;
     @XmlAttribute
     @XmlSchemaType(name = "positiveInteger")
-    private BigInteger width;
+    private Integer width;
     @XmlAttribute
     @XmlSchemaType(name = "positiveInteger")
-    private BigInteger height;
+    private Integer height;
 
     /**
      * An empty constructor used by JAXB.
@@ -78,7 +77,7 @@ public class LogoURL {
      * Build a new LogoURL object.
      */
     public LogoURL(final String format, final OnlineResource onlineResource, 
-            final BigInteger width, final BigInteger height ) {
+            final Integer width, final Integer height ) {
         this.format         = format;
         this.height         = height;
         this.onlineResource = onlineResource;
@@ -104,14 +103,14 @@ public class LogoURL {
     /**
      * Gets the value of the width property.
      */
-    public BigInteger getWidth() {
+    public Integer getWidth() {
         return width;
     }
 
     /**
      * Gets the value of the height property.
      */
-    public BigInteger getHeight() {
+    public Integer getHeight() {
         return height;
     }
 }
